cement_ai_platform/cemet_plant_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.3
pydantic==2.9.0
numpy==1.26.0
pandas==2.1.0
python-multipart==0.0.9
python-dotenv==1.0.0
asyncio==3.4.3
httpx==0.27.0